chore: Add build:watch script for continuous TypeScript compilation
The code changes in the package.json file add a new build:watch script that uses the concurrently package to run TypeScript compilation in watch mode. This script allows for continuous compilation of the TypeScript code, making it easier to develop and test the application without manually running the build command each time.

This commit message follows the established convention of using a prefix to indicate the type of change, followed by a brief description of the change.