chore: Update .env.example with new configuration options
This code change updates the .env.example file to include a new configuration option, LOG. The LOG option is set to "FALSE" by default and can be used to enable or disable logging in the application. This addition improves the configurability of the application and ensures that the default value for the LOG option is documented.

This commit message follows the established convention of using a prefix to indicate the type of change, followed by a brief description of the change.